// auth/auth-service/src/services/proto/userPackage/GoogleGrpcUser.ts
// Original file: src/services/proto/user.proto


export interface GoogleGrpcUser {
  'sub'?: (string);
  'name'?: (string);
  'given_name'?: (string);
  'family_name'?: (string);
  'picture'?: (string);
  'email'?: (string);
  'email_verified'?: (string);
  'locale'?: (string);
}

export interface GoogleGrpcUser__Output {
  'sub': (string);
  'name': (string);
  'given_name': (string);
  'family_name': (string);
  'picture': (string);
  'email': (string);
  'email_verified': (string);
  'locale': (string);
}